requirements.txt
Django==1.4.1
PyYAML==3.10
boto==2.6.0
dj-database-url==0.2.1
django-storages==1.1.5
gunicorn==0.14.6
psycopg2==2.4.5
requests==0.14.1
wsgiref==0.1.2
